+++ back-end/h5-api/config/server.js
@@ -1,0 +1,9 @@
+module.exports = ({ env }) => ({
+  host: env('HOST', '0.0.0.0'),
+  port: env.int('PORT', 1337),
+  admin: {
+    auth: {
+      secret: env('ADMIN_JWT_SECRET', '4f8b422ea4ae379419a2b1c264e79ece'),
+    },
+  },
+});