+++ front-end/h5/src/components/core/editor/modals/preview.vue
@@ -1,6 +1,6 @@
 <script>
 import { mapActions } from 'vuex'
-import QRCode from 'qrcode'
+import ShareInfo from './share-info'
 
 export default {
   props: {
@@ -21,22 +21,13 @@ export default {
     // ...mapState('editor', {
     //   work: state => state.work
     // }),
-    releaseUrl () {
+    previewUrl () {
       return `${window.location.origin}/works/preview/${this.work.id}?view_mode=preview`
     }
   },
-  data () {
-    return {
-      confirmLoading: false,
-      qrcodeSize: 500
-    }
-  },
-  watch: {
-    visible (val) {
-      if (!val) return
-      this.$nextTick(() => this.drawQRcode())
-    }
-  },
+  data: () => ({
+    confirmLoading: false
+  }),
   methods: {
     ...mapActions('editor', [
       'saveWork',
@@ -48,34 +39,16 @@ export default {
         this.handleClose()
         this.confirmLoading = false
       })
-      // setTimeout(() => {
-      // }, 2000);
     },
     handleCancel (e) {
       console.log('Clicked cancel button')
       this.handleClose()
-    },
-    drawQRcode () {
-      var canvas = document.getElementById('qrcode-container')
-      QRCode.toCanvas(canvas, this.releaseUrl, { scale: 4 }, err => {
-        console.log(err)
-      })
     },
     postMessage2Iframe (message) {
       let iframe = document.getElementById('iframe-for-preview')
       if (!iframe) return
       const iframeWin = iframe.contentWindow
       iframeWin.postMessage(message, window.location.origin)
-    },
-    openNewTab (urlType) {
-      switch (urlType) {
-        case 'openPreviewPage':
-          window.open(this.releaseUrl)
-          break
-        case 'buildEngineDocs':
-          window.open('https://ly525.github.io/luban-h5/zh/getting-started/quick-start.html#_2-%E6%9E%84%E5%BB%BA%E9%A2%84%E8%A7%88%E6%89%80%E9%9C%80%E7%9A%84%E6%B8%B2%E6%9F%93%E5%BC%95%E6%93%8E')
-          break
-      }
     }
   },
   render (h) {
@@ -108,7 +81,7 @@ export default {
                     // similar with v-if="this.visible": destory the iframe after close the preview dialog to avoid playing the music and video
                     this.visible && <iframe
                       id="iframe-for-preview"
-                      src={this.releaseUrl}
+                      src={this.previewUrl}
                       frameborder="0"
                       style="height: 100%;width: 100%;"
                     ></iframe>
@@ -118,46 +91,7 @@ export default {
               </div>
             </a-col>
             <a-col span={12} offset={4}>
-              <div class="setting">
-                <div class="info">
-                  <h4 class="label">设置作品信息</h4>
-                  <a-input
-                    class="input"
-                    value={this.work.title}
-                    onChange={e => this.updateWork({ title: e.target.value })}
-                    // onBlur={this.saveTitle}
-                    placeholder="请输入标题"
-                  ></a-input>
-                  <a-input
-                    class="input"
-                    value={this.work.description}
-                    onChange={e => this.updateWork({ description: e.target.value })}
-                    // v-model="description"
-                    // onBlur={this.saveDescription}
-                    placeholder="请输入描述"
-                    type="textarea"
-                  ></a-input>
-                </div>
-                <div class="qrcode my-4">
-                  <div class="label">
-                    <span>手机扫码分享给好友</span>
-                  </div>
-                  <div class="code">
-                    <canvas style="float: left" id="qrcode-container"></canvas>
-                    {/**
-                    <a-radio-group class="radios" value={this.qrcodeSize} onChange={e => { this.qrcodeSize = e.target.value }}>
-                      <a-radio label={500} value={500}>500x500</a-radio>
-                      <a-radio label={1000} value={1000}>1000x1000</a-radio>
-                      <a-radio label={2000} value={2000}>2000x2000</a-radio>
-                    </a-radio-group>
-                    */}
-                  </div>
-                </div>
-                <div style="background: #fafafa;">
-                  <a-button type="link" icon="link" onClick={() => this.openNewTab('openPreviewPage')}>打开预览页面</a-button>
-                  <a-button type="link" icon="link" onClick={() => this.openNewTab('buildEngineDocs')}>如果本地预览显示空白，点此查看文档</a-button>
-                </div>
-              </div>
+              <ShareInfo />
             </a-col>
           </a-row>
         </div>
@@ -217,48 +151,6 @@ export default {
         }
 
       }
-    }
-  }
-  .setting {
-    color: #4a4a4a;
-    font-size: 14px;
-    float: right;
-    width: 380px;
-    .info {
-      .input {
-        margin-top: 10px;
-      }
-    }
-    .qrcode {
-      margin-top: 20px;
-
-      .label span {
-        margin-right: 10px;
-      }
-    }
-    .code {
-      // !#zh 防止浮动塌陷
-      overflow: hidden;
-      .radios {
-        width: 80px;
-        margin-top: 5px;
-        margin-left: 30px;
-        label {
-          margin-left: 0px;
-          margin-top: 10px;
-        }
-        button {
-          margin-top: 15px;
-        }
-      }
-    }
-    .link {
-      width: 100%;
-      display: block;
-    }
-    .edit {
-      text-align: center;
-      margin-top: 20px;
     }
   }
 }