+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.16.0](https://github.com/ly525/luban-h5/compare/v1.15.0...v1.16.0) (2021-03-21)
+
+
+### Features
+
+* script&action&data-source ([24201f6](https://github.com/ly525/luban-h5/commit/24201f631106b811dde877fbcd4cff3fba850068))
+* **component:** tinymce-editor ([0a10d92](https://github.com/ly525/luban-h5/commit/0a10d9261a083ea8bb179d98bdda1ac21a03d0e8))
+
+
+### Bug Fixes
+
+* [#253](https://github.com/ly525/luban-h5/issues/253) ([608c4c4](https://github.com/ly525/luban-h5/commit/608c4c463291779f53760e63f365aac9ce329cc6))
+* add id for data-source ([fb06873](https://github.com/ly525/luban-h5/commit/fb068732058476c5f1ae54891bc78b3990181b9c))
+* treat ds.pipe as String ([7a6ec09](https://github.com/ly525/luban-h5/commit/7a6ec094a9e6f1e6fcdfba26f30122f26a1d24f8))
+
+## [1.15.0](https://github.com/ly525/luban-h5/compare/v1.14.1...v1.15.0) (2021-03-21)
+
+
+### Features
+
+* **canvas:** adjust canvas width; #!zh: 支持调整画布宽度 ([7fe6408](https://github.com/ly525/luban-h5/commit/7fe6408f75e87bad6dc7a48fcf5e78167cd88ac9))
+* remove delete from template card ([d08ebdf](https://github.com/ly525/luban-h5/commit/d08ebdffdf7e2d4d89ec126053c402daf793c7af))
+* 支持制作海报(导出静态图、封面图) !#en: support download poster; [#135](https://github.com/ly525/luban-h5/issues/135) ([929ba0f](https://github.com/ly525/luban-h5/commit/929ba0fd1e84f18b38dd0eb36cfbb00a641b2283))
+
+
+### Bug Fixes
+
+* [#233](https://github.com/ly525/luban-h5/issues/233) add confirm before delete element; #!zh: 删除元素前需要确认 ([349d12e](https://github.com/ly525/luban-h5/commit/349d12efb666143faa463867083ce0a9eab0119b))
+* [#251](https://github.com/ly525/luban-h5/issues/251) ([c2233d2](https://github.com/ly525/luban-h5/commit/c2233d2239973c0cf5ee49f3446380a8377f84b8))
+* preview margin&padding; #!zh:预览margin&padding ([7b14d7c](https://github.com/ly525/luban-h5/commit/7b14d7c26952c386ef303852063235fe32b12243))
+
 ### [1.14.1](https://github.com/ly525/luban-h5/compare/v1.14.0...v1.14.1) (2021-02-13)